requirements.txt
numpy
pyyaml
python-mnist